refactor(site): decompose agents pages so each route owns its data
Move data fetching out of the centralized AgentsPage layout and into
each child route component:

- AgentCreatePage: fetches model catalog/configs, owns create mutation
- AgentDetail: fetches model catalog/configs/desktop flag internally
- AgentsSettingsRoute: computes isAgentsAdmin from useAuthenticated()
- AgentsAnalyticsRoute: already self-contained

This shrinks AgentsOutletContext from 19 fields to 9, keeping only
cross-cutting concerns (chat error tracking, archive actions, sidebar
state). React-query deduplicates the model queries that are now
subscribed to by multiple components.

AgentsPage (the layout) retains model queries for the sidebar, the
chat list, WebSocket watcher, and archive/unarchive mutations.
