+++ site/src/pages/AgentsPage/AgentDetail.stories.tsx
@@ -51,56 +51,7 @@ const AgentDetailLayout: FC = () => {
 							requestUnarchiveAgent: () => {},
 							isSidebarCollapsed: false,
 							onToggleSidebarCollapsed: () => {},
-							modelOptions: [
-								{
-									id: "openai:gpt-4o",
-									provider: "openai",
-									model: "gpt-4o",
-									displayName: "GPT-4o",
-								},
-							],
-							modelConfigIDByModelID: new Map([["openai:gpt-4o", "config-1"]]),
-							modelIDByConfigID: new Map([["config-1", "openai:gpt-4o"]]),
-							modelConfigs: [
-								{
-									id: "config-1",
-									provider: "openai",
-									model: "gpt-4o",
-									display_name: "GPT-4o",
-									enabled: true,
-									is_default: false,
-									context_limit: 200000,
-									compression_threshold: 70,
-									created_at: "2026-01-01T00:00:00Z",
-									updated_at: "2026-01-01T00:00:00Z",
-								},
-							],
-							modelCatalog: {
-								providers: [
-									{
-										provider: "openai",
-										available: true,
-										models: [
-											{
-												id: "openai:gpt-4o",
-												provider: "openai",
-												model: "gpt-4o",
-												display_name: "GPT-4o",
-											},
-										],
-									},
-								],
-							},
-							isModelCatalogLoading: false,
-							modelCatalogError: null,
-							desktopEnabled: false,
-							onCreateChat: () => Promise.resolve(),
-							isCreating: false,
-							createError: undefined,
-							isModelConfigsLoading: false,
-							logoUrl: "",
 							onExpandSidebar: () => {},
-							isAgentsAdmin: false,
 						} satisfies AgentsOutletContext
 					}
 				/>
